restore uploadProfileImage method in UserService
