+++ docker/docker-compose.yml
@@ -3,7 +3,7 @@ version: "3.8"
 services:
   immich-server:
     image: altran1502/immich-server:latest
-    entrypoint: ["/bin/sh", "./start-server.sh"]
+    entrypoint: [ "/bin/sh", "./start-server.sh" ]
     expose:
       - "3000"
     volumes:
@@ -21,7 +21,7 @@ services:
 
   immich-microservices:
     image: altran1502/immich-server:latest
-    entrypoint: ["/bin/sh", "./start-microservices.sh"]
+    entrypoint: [ "/bin/sh", "./start-microservices.sh" ]
     volumes:
       - ${UPLOAD_LOCATION}:/usr/src/app/upload
     env_file:
@@ -37,7 +37,7 @@ services:
 
   immich-machine-learning:
     image: altran1502/immich-machine-learning:latest
-    entrypoint: ["/bin/sh", "./entrypoint.sh"]
+    entrypoint: [ "/bin/sh", "./entrypoint.sh" ]
     expose:
       - "3001"
     volumes:
@@ -54,7 +54,7 @@ services:
 
   immich-web:
     image: altran1502/immich-web:latest
-    entrypoint: ["/bin/sh", "./entrypoint.sh"]
+    entrypoint: [ "/bin/sh", "./entrypoint.sh" ]
     env_file:
       - .env
     ports:
